# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ouranos platform API",
    "version": "1.0.0",
    "description": "Travel platform skeleton. Supabase bearer authentication, organization isolation, versioned commands, synchronization and immutable approval revisions. Planning/voucher form validators and external providers require integration. A successful Ouranos command is not DTS acceptance."
  },
  "servers": [
    {
      "url": "http://localhost:4100",
      "description": "Local development; replace with your configured API origin."
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/openapi.json": {
      "get": {
        "operationId": "openapi",
        "summary": "This generated OpenAPI document",
        "security": [],
        "responses": {
          "200": {
            "description": "OpenAPI 3.0 document",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/session": {
      "get": {
        "operationId": "session",
        "summary": "Verified user and active organization memberships",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "401": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/organizations": {
      "post": {
        "operationId": "createOrganization",
        "summary": "Create an organization with the caller as administrator",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedOrganization"
                }
              }
            }
          },
          "401": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationInput"
              }
            }
          }
        }
      }
    },
    "/v1/organizations/{id}/members": {
      "put": {
        "operationId": "setMembership",
        "summary": "Administrator assigns an existing user a role",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipUpdated"
                }
              }
            }
          },
          "401": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "description": "The caller cannot change their own membership. The target user must already exist in Supabase Auth.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MembershipInput"
              }
            }
          }
        }
      }
    },
    "/v1/commands": {
      "post": {
        "operationId": "executeCommand",
        "summary": "Execute one versioned, idempotent command",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Command rejected. Domain failures include commandId and ok:false; request/database failures use the error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/CommandFailure"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Command rejected. Domain failures include commandId and ok:false; request/database failures use the error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/CommandFailure"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Command rejected. Domain failures include commandId and ok:false; request/database failures use the error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/CommandFailure"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "description": "Use commandId in the body as the idempotency key; the Idempotency-Key HTTP header is not read. Reuse the entire unchanged command when retrying a lost response. expectedVersion is 0 for creation and the current entity version for updates. A deviceId belongs to one user in one organization. Custom date/order, duplicate-expense and final-approver refinements are enforced by Zod at runtime in addition to the generated JSON Schema.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command"
              }
            }
          }
        }
      }
    },
    "/v1/sync/push": {
      "post": {
        "operationId": "pushCommands",
        "summary": "Execute an ordered batch of up to 50 commands",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncPushResult"
                }
              }
            }
          },
          "401": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "description": "Commands commit independently; inspect every result. This is not an all-or-nothing transaction. A request-level failure may occur after earlier commands committed; retry identical envelopes.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncPushInput"
              }
            }
          }
        }
      }
    },
    "/v1/sync/pull": {
      "get": {
        "operationId": "pullChanges",
        "summary": "Read visible changes after a cursor",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncPage"
                }
              }
            }
          },
          "401": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "description": "Cursor values are decimal strings, not JavaScript numbers. Visibility is evaluated for the authenticated user. Use bootstrap to reconcile records that are no longer visible.",
        "parameters": [
          {
            "name": "organizationId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^\\d+$",
              "default": "0"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 100
            }
          }
        ]
      }
    },
    "/v1/sync/bootstrap": {
      "get": {
        "operationId": "bootstrap",
        "summary": "Load the currently visible workspace snapshot",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bootstrap"
                }
              }
            }
          },
          "401": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "description": "Returns at most 2,000 records per entity kind; a larger visible set returns DEPENDENCY_PENDING. Replace server-confirmed cached records absent from this snapshot while retaining pending local drafts for explicit resolution.",
        "parameters": [
          {
            "name": "organizationId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/v1/entities/{kind}": {
      "get": {
        "operationId": "listEntities",
        "summary": "List visible entities of one kind",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityList"
                }
              }
            }
          },
          "401": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "description": "Sorted by updated_at descending, then id. limit is supported; cursor is currently ignored for this route. Use sync/pull for cursor-based change retrieval.",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EntityKind"
            }
          },
          {
            "name": "organizationId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 100
            }
          }
        ]
      }
    },
    "/v1/entities/{kind}/{id}": {
      "get": {
        "operationId": "getEntity",
        "summary": "Load one visible entity",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityResponse"
                }
              }
            }
          },
          "401": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Request could not be completed. Inspect error.code; error.details may be absent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EntityKind"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "organizationId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Supabase access token"
      }
    },
    "schemas": {
...
      "EntityKind": {
        "type": "string",
        "enum": [
          "trip",
          "authorization",
          "expense",
          "voucher",
          "document",
          "approval",
          "workflow",
          "notification",
          "integration"
        ]
      },
...
      "OrganizationInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          }
        },
        "required": [
          "name"
        ],
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "VALIDATION_FAILED",
                  "VERSION_CONFLICT",
                  "PERMISSION_DENIED",
                  "AUTHENTICATION_REQUIRED",
                  "NOT_FOUND",
                  "ATTACHMENT_INCOMPLETE",
                  "INVALID_STATE_TRANSITION",
                  "IDEMPOTENCY_CONFLICT",
                  "DEPENDENCY_PENDING",
                  "PROVIDER_UNAVAILABLE",
                  "RATE_LIMITED",
                  "INTERNAL_ERROR"
                ]
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "CreatedOrganization": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "additionalProperties": false
      },
      "MembershipUpdated": {
        "type": "object",
        "properties": {
          "updated": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "updated"
        ],
        "additionalProperties": false
      },
...
    }
  }
}